Implemented basic player movment
Created and implemented basic player movement. Using the "LocomotionHandler" script, I pulled the move direction from the "PlayerInput" script. Then I added velocity to the player based on the current input movement direction. Finally I clamped the velocity value to the maximum speed so that the player does not fly out of control. This mechanic will be developed further in future commits.