Implemented prototype Health System mechanic
A Health System mechanic has been developed and implemented, which utilizes a Health Handler script that is exclusively enabled during Space Exploration. Once activated, the script adjusts the player's health to the maximum health value, and subsequently decreases the health by a depletion rate. It should be noted that this is a prototype and will undergo further development in future commits.