Updated Object Emission Manager
Updated the Object Emission Manager to always be positioned on the player and spawn objects at random points around the player off screen. The object spawn at an interval set by editor settings. The interval time decreases over time until player is dead. This is a prototype and will updated in future commits.