Implemented High Score System
Created and implemented a High Score System that uses a HighScoreHandler script to track the players high score, load high score, and save high score if new high score is achieved. The high score is saved to player prefs but with use custom save mechanic in the future. This is a prototype and will be revised in later commits.