Implemented Player Size system
A Player Size System has been developed and implemented, which involves the use of the PlayerManager script to store relevant functions that can be called upon by multiple scripts. One of these scripts is the ObjectHandler script, which triggers the "increase player size" function to enlarge the player by a playerIncreaseRate variable. This variable can be adjusted from the editor. It is important to note that this is currently a prototype and will be subject to further improvements in future updates.