Implemented a Camera Manager
Developed and integrated a Camera Manager singleton, which comprises a CameraManager script and a CinemachineBrain component. The Camera Manager searches for the Cinemachine camera with the highest priority in the scene and sets it as the main camera. Additionally, the Camera Manager script contains functions that enable manipulation of the camera as needed.