Prototyped new VFX for obstacles and implemented Reset Score button
I created a prototype of the obstacle object's VFX, which will require additional refinement in subsequent commits. Additionally, I added a "Reset Score" button to the Title Menu that enables players to reset their high scores. It is important to note that these mechanics are currently in their prototype stage and will be improved upon in future commits.