Replaced Ship Hub with Main Menu
Prototyped and implemented a Main Menu scene to replace the Ship Hub scene. This was done to save time and cut unnecessary mechanics. The Main Menu will still act like a ship hub but will not allow the player to move and interact organically with it. This is a prototype and will be revised in future updates.